Major Tidy up
I have bundled the majority of the consumer loop so that it fits more nicely and I have also erased the TLPM photodetector non library and replaced it with the library files that are already installed (the ones I have developped) in nxg. Also this is Labview 5.0.
I have compiled succesfully the executable.